Refactored `put()` method to allow for "restartable uploads"
...otherwise large files may fail to upload. The problem is that file
might be located at some (slave) node, i.e., on a different node the
code running. Therefore we have to first copy contents of the file
to the local filesystem and then let WebDAV library to access that file
(so it can seek freely and restart / split the upload if necessary).

Hope it helps.